feat: backend for statistics. changed time precision to ms from s

+++ src/domain/typing_test_input.h
@@ -1,9 +1,12 @@
 #ifndef TYPING_TEST_INPUT_H
 #define TYPING_TEST_INPUT_H
 
+#include <stdbool.h>
+#include <stdint.h>
+
 typedef struct TypingTestInput {
     bool is_correct;
-    int  time_since_test_start;
+    int64_t time_since_test_start;
     // has to be an int since there are other non-alphabetic keycodes.
     int inputted;
 } TypingTestInput;
